Fix tool
